Renewal request feedback updates, feature gate
Only allows creating one request per loan, even if previous approved.
Clarifies some functionality and reorganizes layout based on feedback.
Sets up feature gating based on environment variables
