+++ .env.development
@@ -29,3 +29,5 @@ BUCKETEER_AWS_ACCESS_KEY_ID=
 BUCKETEER_AWS_REGION=
 BUCKETEER_AWS_SECRET_ACCESS_KEY=
 BUCKETEER_BUCKET_NAME=
+
+FEATURE_RENEWAL_REQUESTS=on
