+++ .env.test
@@ -3,3 +3,5 @@ SQUARE_LOCATION_ID=SQLOCID1234
 SQUARE_ENVIRONMENT=sandbox
 
 APPOINTMENT_SLOT_CALENDAR_ID=appointmentSlots@calendar.google.com
+
+FEATURE_RENEWAL_REQUESTS=on
