+++ test/controllers/admin/renewal_requests_controller_test.rb
@@ -1,0 +1,6 @@
+require "test_helper"
+
+module Admin
+  class RenewalRequestsControllerTest < ActionDispatch::IntegrationTest
+  end
+end
